RunResult: Move yaml tools
