+++ pisek/jobs/parts/run_result.py
@@ -10,14 +10,17 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 from enum import Enum
-from typing import Optional, Union, Callable
+from typing import Optional, Union
+import yaml
 
 from pisek.utils.paths import TaskPath
+
 
 class RunResultKind(Enum):
     OK = 0
     RUNTIME_ERROR = 1
     TIMEOUT = 2
+
 
 class RunResult:
     """Represents the way the program execution ended. Specially, a program
@@ -42,3 +45,36 @@ class RunResult:
         self.time = time
         self.wall_time = wall_time
 
+
+def run_result_representer(dumper, run_result: RunResult):
+    return dumper.represent_sequence(
+        "!RunResult",
+        [
+            run_result.kind.name,
+            run_result.returncode,
+            run_result.time,
+            run_result.wall_time,
+            run_result.stdout_file,
+            run_result.stderr_file,
+            run_result.status,
+        ],
+    )
+
+
+def run_result_constructor(loader, value):
+    (
+        kind,
+        returncode,
+        time,
+        wall_time,
+        out_f,
+        err_f,
+        status,
+    ) = loader.construct_sequence(value)
+    return RunResult(
+        RunResultKind[kind], returncode, time, wall_time, out_f, err_f, status
+    )
+
+
+yaml.add_representer(RunResult, run_result_representer)
+yaml.add_constructor("!RunResult", run_result_constructor)
